Changed requirements for using Enzyme in README.md
The existing tutorial triggered errors and warnings.

Starting with React 15.5 the requirements for Enzyme changed. The up to date installation instructions can be seen here: https://github.com/airbnb/enzyme#installation

I have opted to leave the instructions for `React <15.5` as a note in parentheses and not write both bash commands because users following this tutorial are instructed to first install Create React App and then use it to generate a new React app. In other words, most readers will have the latest React version.