Add instructions for existing API
Finding out this information wasted 2-3 hours of dev time for me today.  We need to serve the files from a custom Hapi server, and it's not obvious that this project can actually be created as a subproject of an existing server.

I think this info will help a lot of people!  Thanks for your consideration.