Fix README instruction on installing enzyme
According to the Enzyme docs, the installation procedure should include the enzyme-adapter-react-16 package as well. Its declaration file is need too, because TS.
Tests will fail for people following the step-by-step instructions unless these packages are added. This could close #69 and #76.
Also changing the test template to load the adapter.