Update tests in README to use enzyme.dive()
When following the instructions in the readme i wasn't able to get the tests working. After some playing around it seems .dive is required when shallow rendering components with Enzyme. Otherwise render is never called on the component.

I don't have much React/Enzyme knowledge so i'm not sure if this is just on my end. But I don't think that is the case here.